+++ CHANGELOG.md
@@ -29,12 +29,15 @@ const setBookmarkExpanded = ({ bookmark, depth, doc, index }) => true;
 const setBookmarkExpanded = ({ bookmark, depth, doc, index }) => false;
 ```
 
+-   The Page Navigation plugin provides new `NumberOfPages` component that displays the total number of pages
+
 **Improvements**
 
 -   Align bookmark titles
 -   Compatible with React 18
 -   In previous versions, all bookmarks were expanded by default. From this version, a bookmark will be shown or hidden initially depending on its data structure.
     You can see this behaviour on popular viewers such as Acrobat Reader.
+-   The toolbar slot `NumberOfPages` provides the ability of customizing the number of pages
 
 ## v3.2.0